Attempt to squelch a narrowing warning.
Some older versions of Clang appear to think dividing a float by a
64-bit integer is narrowing 64 bits to 32 bits.

Reviewed-on: https://code.wireshark.org/review/29518

+++ ui/io_graph_item.c
@@ -204,7 +204,7 @@ double get_io_graph_item(const io_graph_item_t *items_, io_graph_item_unit_t val
             break;
         case IOG_ITEM_UNIT_CALC_AVERAGE:
             if (item->fields) {
-                value = item->float_tot / item->fields;
+                value = (double)item->float_tot / item->fields;
             } else {
                 value = 0;
             }
